pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "warp-xpbd-neo-hook"
version = "0.0.1"
requires-python = ">=3.9"
dependencies = ["numpy", "warp-lang"]

[tool.setuptools.packages.find]
where = ["."]
